+++ practica-3/lib/binary.c
