+++ practica-3/lib/binary.h
